app2/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
langchain
langchain-core
langchain-google-genai
langchain-huggingface
langchain-community
sentence-transformers
faiss-cpu
google-cloud-storage
requests